feature: create a home page component and separate all components by page type
